Modifications to suppress warnings, suggested by C.Pulvermacher
git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@6118 b38c6c8f-bee8-465d-8448-9abe6915d496

+++ trg/cdc/src/Hough3DFinder.cc
@@ -505,7 +505,7 @@ namespace Belle2 {
       double mcZ0 = trackMCParticle.getVertex().Z()/100;
       double mcCot = trackMCParticle.getMomentum().Pz()/trackMCParticle.getMomentum().Pt();
       // Set input of finder
-      double tempTrackVariables[] = {charge, rho, fitPhi0};
+      double tempTrackVariables[] = {double(charge), rho, fitPhi0};
       vector<double > trackVariables(tempTrackVariables, tempTrackVariables+sizeof(tempTrackVariables) / sizeof(tempTrackVariables[0]) );
 
       // Run finder
